Inventory Update Main.java
Finally finished the Inventory page stuff. Rows an be added, deleted, saved, and are automatically loaded from "inventory.txt". No capitals in the name.

So we'll need a file like that. I'm still new to Github, so if I haven't figured it out by the time you're reading this, M0th, do me a favour with that?